requirements.txt
anyio==4.3.0
blinker==1.7.0
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Script==2.0.6
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
keycloak==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
urllib3==2.2.1
Werkzeug==3.0.2
